+++ src/navigation/AuthedNavigator.tsx
@@ -1,90 +1,39 @@
-import { createBottomTabNavigator } from '@react-navigation/bottom-tabs'
-import { Home, List, MessageCircle, Settings } from '@tamagui/lucide-icons'
-import { HomeFeed } from 'app/views/feed/HomeFeed'
-import { useNostr } from 'lib/hooks'
-import { useEffect } from 'react'
-import { XStack } from 'tamagui'
-import { NavHeader } from 'views/shared'
-import { SettingsScreen } from 'views/user/SettingsScreen'
+import { createNativeStackNavigator } from '@react-navigation/native-stack'
+import { ProfileScreen } from 'views/profile'
 
-import { ChatNavigator } from './ChatNavigator'
+import { TabNavigator } from './TabNavigator'
 
-const BottomTab = createBottomTabNavigator<{
-  feed: undefined
-  chat: undefined
-  settings: undefined
-}>()
+export type AuthedStackParams = {
+  tabs: undefined
+  profile: { pubkey: string }
+}
 
-const activeTabColor = '$color12'
-const inactiveTabColor = '$color8'
+const Stack = createNativeStackNavigator<AuthedStackParams>()
 
 export function AuthedNavigator() {
-  const nostr = useNostr()
-  useEffect(() => {
-    if (!nostr) return
-    nostr.setupInitialSubscriptions()
-  }, [nostr])
   return (
-    <BottomTab.Navigator
-      initialRouteName="feed"
+    <Stack.Navigator
+      initialRouteName="tabs"
       screenOptions={{
-        headerShown: false,
-        tabBarLabel: () => null,
-        tabBarStyle: {
-          borderTopWidth: 0,
+        headerStyle: {
+          backgroundColor: '#f4511e',
         },
-        tabBarBackground: () => (
-          <XStack
-            f={1}
-            backgroundColor="$backgroundSoft"
-            borderTopWidth="$1"
-            borderTopColor="$color4"
-            elevation="$6"
-          />
-        ),
+        headerTintColor: '#fff',
+        headerTitleStyle: {
+          fontWeight: 'bold',
+        },
       }}
     >
-      <BottomTab.Screen
-        name="feed"
-        component={HomeFeed}
-        options={{
-          tabBarIcon: ({ focused, size }) => (
-            <Home
-              color={focused ? activeTabColor : inactiveTabColor}
-              size={size}
-            />
-          ),
-        }}
+      <Stack.Screen
+        name="tabs"
+        component={TabNavigator}
+        options={{ title: 'Chats' }}
       />
-      <BottomTab.Screen
-        name="chat"
-        component={ChatNavigator}
-        options={{
-          tabBarIcon: ({ focused, size }) => (
-            <MessageCircle
-              color={focused ? activeTabColor : inactiveTabColor}
-              size={size}
-            />
-          ),
-        }}
+      <Stack.Screen
+        name="profile"
+        component={ProfileScreen}
+        options={{ title: 'Chat' }}
       />
-      <BottomTab.Screen
-        name="settings"
-        component={SettingsScreen}
-        options={{
-          headerShown: true,
-          tabBarIcon: ({ focused, size }) => (
-            <Settings
-              color={focused ? activeTabColor : inactiveTabColor}
-              size={size}
-            />
-          ),
-          title: 'Settings',
-          header: ({ options }) => (
-            <NavHeader options={options} title={options.title} />
-          ),
-        }}
-      />
-    </BottomTab.Navigator>
+    </Stack.Navigator>
   )
 }
